saving work for now

+++ VortexEngine/src/Wireless/VLReceiver.h
@@ -71,10 +71,10 @@ private:
   static uint8_t m_pinState;
 
   // used to compare if received data has changed since last checking
-  static uint32_t m_previousBytes;
+  static uint16_t m_previousBytes;
 
   // the determined time based on sync
-  static uint32_t m_vlTiming;
+  static uint16_t m_vlTiming;
 
   // count of the sync bits (similar length starter bits)
   static uint8_t m_syncCount;
